Fix: check that dir exists before trying to remove it

+++ app/icon_generation.php
@@ -621,16 +621,18 @@ function hsl2rgb($h, $s, $v) {
  *
  */
 function _delete_dir($dir_path) {
-  $it = new RecursiveDirectoryIterator($dir_path, RecursiveDirectoryIterator::SKIP_DOTS);
-  $files = new RecursiveIteratorIterator($it,
-    RecursiveIteratorIterator::CHILD_FIRST);
-  foreach ($files as $file) {
-    if ($file->isDir()) {
-      rmdir($file->getRealPath());
+  if (is_dir($dir_path)) {
+    $it = new RecursiveDirectoryIterator($dir_path, RecursiveDirectoryIterator::SKIP_DOTS);
+    $files = new RecursiveIteratorIterator($it,
+      RecursiveIteratorIterator::CHILD_FIRST);
+    foreach ($files as $file) {
+      if ($file->isDir()) {
+        rmdir($file->getRealPath());
+      }
+      else {
+        unlink($file->getRealPath());
+      }
     }
-    else {
-      unlink($file->getRealPath());
-    }
+    rmdir($dir_path);
   }
-  rmdir($dir_path);
 }
